restore monitor mode
This commit aims to address the undesired behaviour where the user
runs an OVE command and discovers that the monitor mode has been
disabled but not restored.

This commit adds "set -m" calls to two functions where the monitor
mode was disabled, but the re-enablement of the monitor mode was
forgotten.

Furthermore, this commit makes sure that the monitor mode is always
restored to its previous value when the ove.bash wrapper is used.